implemented enum class "TypeOfUser" for user
