--- Project-SpringBoot-DevelhopeMusic/src/main/java/it/develhope/develhopeMusic/entities/User.java
@@ -1,62 +1,0 @@
-package it.develhope.develhopeMusic.entities;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "users")
-public class User {
-    private int id;
-    private String username;
-    private String password;
-    private String email;
-    private String typeOfUser;
-
-    public User() {
-    }
-
-    public User(int id, String username, String password, String email, String typeOfUser) {
-        this.id = id;
-        this.username = username;
-        this.password = password;
-        this.email = email;
-        this.typeOfUser = typeOfUser;
-    }
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getTypeOfUser() {return typeOfUser;}
-
-    public void setTypeOfUser(String typeOfUser) {this.typeOfUser = typeOfUser;}
-}
