+++ Project-SpringBoot-DevelhopeMusic/src/main/java/it/develhope/develhopeMusic/entities/user/TypeOfUser.java
@@ -1,0 +1,16 @@
+package it.develhope.develhopeMusic.entities.user;
+
+public enum TypeOfUser {
+
+    LISTENER {
+        public String toString() {
+            return "LISTENER";
+        }
+    },
+
+    ARTIST {
+        public String toString() {
+            return "ARTIST";
+        }
+    }
+}
