+++ Project-SpringBoot-DevelhopeMusic/src/main/java/it/develhope/develhopeMusic/repositories/UserRepository.java
@@ -1,6 +1,6 @@
 package it.develhope.develhopeMusic.repositories;
 
-import it.develhope.develhopeMusic.entities.User;
+import it.develhope.develhopeMusic.entities.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
